Add files via upload
The project aims at predicting customers' reorder based on their previous order history. It gives an insight on which day of the month, on which day of the week and on which hour of the day the store observes the maximum number of orders. Also, it gives an insight into which category of items is most popular among reorders.